[ style ] 컴포넌트 분리 및 컴포넌트 이름 변경
1. MovieInfo -> MovieArticle로 이름 변경
2. MovieInfo라는 또 다른 세부 컴포넌트 생성
3. 컴포넌트 분리
